Refactor repeated pattern to a helper function
