+++ integration_tests/features/ValidatorNode.feature
@@ -13,7 +13,7 @@ Feature: Validator Node
         And I mine 4 blocks using wallet WALLET1 on NODE1
         And I have a validator node VN1 connected to base node NODE1 and wallet WALLET1
         When I publish a contract acceptance transaction for contract DEF1 for the validator node VN1
-        And I mine 4 blocks using wallet WALLET1 on NODE1
+        And I mine 9 blocks using wallet WALLET1 on NODE1
         Then wallet WALLET1 will have a successfully mined contract acceptance transaction for contract DEF1
 
     @dan
@@ -30,19 +30,18 @@ Feature: Validator Node
         When I create a contract constitution COM1 for contract DEF1 from file "fixtures/contract_constitution.json"
         And I add VN1 to the validator committee on COM1
         And I publish the contract constitution COM1 on wallet WALLET1 via command line
-        And I mine 4 blocks using wallet WALLET1 on NODE1
+        And I mine 9 blocks using wallet WALLET1 on NODE1
         Then wallet WALLET1 will have a successfully mined constitution acceptance transaction for contract DEF1
 
     @dan @critical
     Scenario: Publish contract update proposal acceptance
         Given I have a seed node NODE1
         And I have wallet WALLET1 connected to all seed nodes
-        When I mine 9 blocks using wallet WALLET1 on NODE1
-        Then I wait for wallet WALLET1 to have at least 1000000 uT
+        And I mine 9 blocks using wallet WALLET1 on NODE1
+        And I wait for wallet WALLET1 to have at least 1000000 uT
         And I publish a contract definition DEF1 from file "fixtures/contract_definition.json" on wallet WALLET1 via command line
-        When I mine 8 blocks using wallet WALLET1 on NODE1
-        Then wallet WALLET1 has at least 1 transactions that are all TRANSACTION_STATUS_MINED_CONFIRMED and not cancelled
+        And I mine 4 blocks using wallet WALLET1 on NODE1
         And I have a validator node VN1 connected to base node NODE1 and wallet WALLET1
-        Then I publish a contract update proposal acceptance transaction for the validator node VN1
-        When I mine 8 blocks using wallet WALLET1 on NODE1
-        Then wallet WALLET1 has at least 2 transactions that are all TRANSACTION_STATUS_MINED_CONFIRMED and not cancelled
+        When I publish a contract update proposal acceptance transaction for the validator node VN1
+        And I mine 9 blocks using wallet WALLET1 on NODE1
+        Then wallet WALLET1 will have a successfully mined contract update proposal for contract DEF1
